// src/store/main/system/types.ts
/*
 * @Author: your name
 * @Date: 2021-08-15 20:44:26
 * @LastEditTime: 2021-08-18 23:30:07
 * @LastEditors: Please set LastEditors
 * @Description: In User Settings Edit
 * @FilePath: \vue3-ts-cms\src\store\main\system\types.ts
 */
export interface ISystemState {
  usersList: any[]
  usersCount: number
  roleList: any[]
  roleCount: number
}